Create json-stringify-raw
Starting with my Node.js project template, create an implementation of
`JSON.stringify` which provides the behavior specified in ES2019, with
the exception that the replacer function must return strings, which are
used verbatim in the JSON, or undefined/null if not replaced.

This addresses a recurring problem with some parsers distinguishing
between integer and non-integer numbers, treating decimal digits as
significant, treating exponential numbers differently, or providing
extensions for NaN/Infinity/Date/RegExp which can not be used from the
standard JSON.stringify.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>