feat: support filtering from replacer
For feature parity with `JSON.stringify`, which omits values where
replacer returns `undefined` (except Array elements, which become null),
support filtering by returning boolean values.

`false` was chosen for filtering instead of `undefined`, which would
have matched `JSON.stringify`, due to the convenience of writing
functions which return `undefined` for values which should not be
replaced, as evidenced by the existing examples which use this
convention.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>